중복삭제 가드문 출력

+++ BookSearcherApp/BookSearcherApp/Scenes/FavoriteList/Reactor/FavoriteListReactor.swift
@@ -17,11 +17,13 @@ final class FavoriteListReactor: BaseReactor<
     // 상태변경 이벤트 정의 (상태를 어떻게 바꿀 것인가)
     enum Mutation {
         case setFavoriteBooks([FavoriteBook]) // 즐겨찾기 세팅 뮤테이션
+        case setDeleting(Bool)
     }
 
     // View의 상태 정의 (현재 View의 상태값)
     struct State {
         var books: [FavoriteBook]
+        var isDeleting: Bool = false
     }
 
     // 생성자에서 초기 상태 설정
@@ -38,12 +40,14 @@ final class FavoriteListReactor: BaseReactor<
             let books = CoreDataManger.shared.fetchAllFavoriteBooks()
             return .just(.setFavoriteBooks(books))
         case let .deleteFavoriteBook(book):
+            guard !currentState.isDeleting else { return .empty() } // 삭제중이면 접근금지 -> 중복삭제 회피
             CoreDataManger.shared.deleteOneFavoriteBook(isbn: book.isbn)
             let books = CoreDataManger.shared.fetchAllFavoriteBooks()
-            return .just(.setFavoriteBooks(books))
+            return .concat(.just(.setDeleting(true)), .just(.setFavoriteBooks(books)), .just(.setDeleting(false)))
         case .clearFavoriteBooks:
+            guard !currentState.isDeleting else { return .empty() } // 삭제중이면 접근금지 -> 중복삭제 회피
             CoreDataManger.shared.deleteAllFavoriteBooks()
-            return .just(.setFavoriteBooks([]))
+            return .concat(.just(.setDeleting(true)), .just(.setFavoriteBooks([])), .just(.setDeleting(false)))
         }
     }
 
@@ -54,6 +58,8 @@ final class FavoriteListReactor: BaseReactor<
         switch mutation {
         case let .setFavoriteBooks(books):
             newState.books = books
+        case let .setDeleting(isDeleting):
+            newState.isDeleting = isDeleting
         }
         return newState
     }
